refactor: 게임기록 관련 기능 분리
- utils 함수로 분리 (util -> utils)
