+++ week3/src/constant/constant.js
@@ -1,4 +1,4 @@
-import { deepFreeze } from '@/util/deepFreeze';
+import { deepFreeze } from '@utils/deepFreeze';
 
 export const MENU_ITEMS = Object.freeze({
   GAME: '게임',
